Modifying ChatController
Modifying ChatController to handle request and send reponse

+++ Backend/ChatServer/chat/src/main/java/chat/repository/ChatRepository.java
@@ -25,7 +25,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ChatRepository extends JpaRepository<Chat, Integer>  {
+public interface ChatRepository extends JpaRepository<Chat, Long>  {
 	
 	List<Chat> findAll();
 }
